+++ algonaut_transaction/src/lib.rs
@@ -2,6 +2,7 @@ pub mod account;
 mod api_model;
 pub mod auction;
 pub mod builder;
+pub mod contract_account;
 pub mod error;
 pub mod transaction;
 pub mod tx_group;
